Add Genkouyoushi Generation
Add rendering functions to the library for genkouyoushi paper with a
configurable layout. The render function assembles the inner grid and
all the page margins. The renderGrid function is responsible for
calculating the sizes of the boxes depending on the available area and
assembling & joining the boxes. The drawBox function is responsible for
generating a single Kanji box with optional sidebox for the Furigana.

The application currently generates a static layout - CLI arguments to
modify the layout & spacing will be added later.