Add Simple API Server for Paper Generation
Add a new `genkouyoushi-api` executable that runs a web server built
with servant and the project's library. This listens at the root URI for
a POST request with JSON for the library's Config type. It replies to
requests with a ByteString of the generated PDF file.

Override the version we use of the diagrams-rasterific package to
incorporate the ability to write PDF files to ByteStrings instead of
files.